refactor: change folders to respect clean arch in auth module
