+++ src/modules/auth/auth.module.ts
@@ -2,10 +2,10 @@ import { Module } from '@nestjs/common';
 import { AuthController } from './infrastructure/auth.controller';
 import { HttpModule } from '@nestjs/axios';
 import { ConfigModule } from '@nestjs/config';
-import { AuthGuard } from './auth.guard';
+import { AuthGuard } from './application/auth.guard';
 import { APP_GUARD } from '@nestjs/core';
 import { JwtModule } from '@nestjs/jwt';
-import { jwtConstants } from './constants/jwt.constants';
+import { jwtConstants } from './application/constants/jwt.constants';
 import { UserModule } from '../user/user.module';
 import { AuthService } from './application/auth.service';
